+++ monitoring.h
@@ -41,6 +41,6 @@ void compare_phs_cpu_vs_ref (FILE *fp, int n_events_val, int n_events_gen,
                              int *channels, int n_in, int n_out, phs_val_t *pval,
                              phs_prt_t *prt, double *factors, double *volumes);
 
-long long required_gpu_mem (int n_events, int n_x);
-long long required_cpu_mem (int n_events, int n_x);
+long long required_gpu_mem (long long n_events, int n_x);
+long long required_cpu_mem (long long n_events, int n_x);
 #endif
